Relocate some of the services tests to test_services
